simplifying the human input
this update addresses the situations in which the auto-solve gets stuck
at a certain point. Here a user must randomly try a ‘possible digit’ in
one of the boxes, and the program tries to solve further.

If it fails, it returns back to this ‘last correct state’ where in the
user can then try the other ‘possible digit/s’ and voila!

have done this by mostly adding functions for ‘saving digits’ and
‘loading digits’ and ‘detecting errors’ which the program does in case
it finds an empty box after basicAnalysis (which should never be
possible)