crudeConversationalSemiFriendlyUI
Made the UI a little conversation and slightly friendly, because why
not, right?
Threw in Count Dooku (an image of him as well). Added a bit of the josh
and retard elements.

I think thats it. Im done now.

Maybe il play with the instruction text and make that a little silly as
well